+++ src/util/Sample.java
@@ -8,17 +8,22 @@ public class Sample {
     @Getter private int index;
     @Getter @Setter private String group;
     @Getter @Setter private String population;
-    @Getter @Setter private String name;
-    @Getter @Setter private ArrayList<Integer> alleleData;
-    @Getter private String alleleType;
+    @Getter
+    @Setter
+    private String name;
+    @Getter
+    @Setter
+    private ArrayList<Float> alleleData;
+    @Getter
+    private String alleleType;
 
-    public Sample(int index,String name,String population, String group, ArrayList<Integer> allaleData) {
+    public Sample(int index, String name, String population, String group, ArrayList<Float> allaleData) {
         super();
         this.index = index;
         this.group = group;
         this.population = population;
         this.name = name;
         this.alleleData = allaleData;
-        this.alleleType=String.join("-",allaleData.toString());
+        this.alleleType = String.join("-", allaleData.toString());
     }
 }
